Initial commit
This is the initial functionality. It supports the default buttons
in the Etherpad formatting toolbar, and uses its own HTML parser
to format the text into wiki markup. Hopefully that will change
over the next week as I figure out how to get EPL to work with me
a little more nicely.

Change-Id: I82feeb2f1b28f367e409e3df7d797183b052d299