Hide other format bars, and delete only after others have left
This is a significant change in the way the backend handles pads,
here's how it works now: Create a group for the page and revision
being edited, then add a pad to it. When someone submits the form
and thereby leaves the pad, we check to see if any sessions are
left in the group. If there are some, we leave the pad. Once the
users are all gone, we delete the pad. This won't affect single
users who use the Etherpad editor, because they'll be the only
people in the pad.

Also, other format bars (default and WikiEditor) are now hidden by
this extension.

Change-Id: I897c9b9fcc9c20e807a284591a398c40abfd032c