Fix revision handling
It doesn't make sense to postfix the revision number to the pad ID,
so let's stop doing that. Also, I moved the code for getting the
pad contents out of the jquery.etherpad library, because there's
no reason for it to call mw functions.

Patchset 2: Fix bug on submit (problematic)

Change-Id: I881cae1845e81d2e0a20fad4f61a5679c808bdec