Add in a better session menu, and various other changes
This patch will make the session menu a little nicer. It will also
slighly improve the user list, and remove the ethereditor-warn
module from the project, since it is no longer used.

It is now working beautifully, though it will depend on another
patch for now, because that patch has some big bugfixes that were
necessary for this one to work.

Also, I've fixed a bug or two along the way.

Change-Id: Ic8ddc344d9974556b38ef8d59b4a6f0e09817da3